+++ src/frontend-react/src/pages/access/MembersTab.tsx
@@ -1,0 +1,3 @@
+export default function MembersTab() {
+  return <p className="text-sm text-muted-foreground">Members tab — coming up.</p>
+}
