+++ src/frontend-react/src/pages/access/OrgsTab.tsx
@@ -1,0 +1,3 @@
+export default function OrgsTab() {
+  return <p className="text-sm text-muted-foreground">Orgs tab — coming up.</p>
+}
